Apply code review requested changes
