+++ tests/test_validators.py
@@ -719,23 +719,6 @@ def test_assert_raises_validation_error():
     ]
 
 
-def test_optional_validator():
-    val_calls = []
-
-    class Model(BaseModel):
-        something: Optional[str]
-
-        @validator('something')
-        def check_something(cls, v):
-            val_calls.append(v)
-            return v
-
-    assert Model().dict() == {'something': None}
-    assert Model(something=None).dict() == {'something': None}
-    assert Model(something='hello').dict() == {'something': 'hello'}
-    assert val_calls == [None, 'hello']
-
-
 def test_required_optional():
     class Model(BaseModel):
         nullable1: Optional[int] = ...
